requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
einops==0.8.1
filelock==3.13.1
frozenlist==1.7.0
fsspec==2024.6.1
ftfy==6.3.1
future==1.0.0
graphviz==0.21
grpcio==1.74.0
hf-xet==1.1.5
huggingface-hub==0.34.3
idna==3.10
jinja2==3.1.4
lightning-utilities==0.15.0
markdown==3.8.2
markupsafe==2.1.5
mpmath==1.3.0
multidict==6.6.3
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
open-clip-torch==2.32.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.0.0
propcache==0.3.2
protobuf==6.31.1
pydeprecate==0.3.1
python-dateutil==2.9.0.post0
pytorch-lightning==1.5.0
pytz==2025.2
pyyaml==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.5.3
six==1.17.0
sympy==1.13.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
timm==1.0.19
tokenizers==0.21.4
# torch==2.7.1+cu118
torchmetrics==1.8.0
# torchvision==0.22.1+cu118
torchviz==0.0.3
tqdm==4.67.1
transformers==4.53.2
triton==3.3.1
typing-extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
werkzeug==3.1.3
xformers==0.0.31.post1
yarl==1.20.1